+++ egui/src/data/input.rs
@@ -152,6 +152,16 @@ pub struct Modifiers {
     pub command: bool,
 }
 
+impl Modifiers {
+    pub fn is_none(&self) -> bool {
+        self == &Self::default()
+    }
+
+    pub fn any(&self) -> bool {
+        !self.is_none()
+    }
+}
+
 /// Keyboard keys.
 ///
 /// Includes all keys egui is interested in (such as `Home` and `End`)
@@ -160,6 +170,7 @@ pub struct Modifiers {
 /// Many keys are omitted because they are not always physical keys (depending on keyboard language), e.g. `;` and `§`,
 /// and are therefor unsuitable as keyboard shortcuts if you want your app to be portable.
 #[derive(Clone, Copy, Debug, Eq, Ord, PartialEq, PartialOrd, Hash)]
+#[cfg_attr(feature = "persistence", derive(serde::Deserialize, serde::Serialize))]
 pub enum Key {
     ArrowDown,
     ArrowLeft,